rename models to include "-" between obvious words
rename models to include "-" between obvious words

rename models to include "-" between obvious words

rename models to include "-" between obvious words
