+++ src/app/models/dto/class-level-sub.model.ts
@@ -1,0 +1,6 @@
+export interface ClassLevelSub {
+  id: number;
+  name: string;
+  class_level_id: number;
+  number_of_applications?: number;
+}
