+++ src/app/models/dto/class-level.model.ts
@@ -1,0 +1,6 @@
+export interface ClassLevel {
+  id: number;
+  name: string;
+  section_id: number;
+  classLevelSubs: ClassLevel[]
+}
