--- src/app/models/dto/classlevel.model.ts
@@ -1,6 +1,0 @@
-export interface ClassLevel {
-  id: number;
-  name: string;
-  section_id: number;
-  classLevelSubs: ClassLevel[]
-}
