--- src/app/models/dto/classlevelsub.model.ts
@@ -1,6 +1,0 @@
-export interface ClassLevelSub {
-  id: number;
-  name: string;
-  class_level_id: number;
-  number_of_applications?: number;
-}
