+++ src/app/models/dto/subject-registration.model.ts
@@ -1,0 +1,8 @@
+export interface SubjectRegistration {
+  id: number;
+  student_id: number;
+  year_id: number;
+  subject_id: number;
+  created_at?: string;
+  updated_at?: string;
+}
