--- src/app/models/dto/subjectregistration.model.ts
@@ -1,8 +1,0 @@
-export interface SubjectRegistration {
-  id: number;
-  student_id: number;
-  year_id: number;
-  subject_id: number;
-  created_at?: string;
-  updated_at?: string;
-}
