+++ src/app/modules/reusable/class-level/save-class/save-class.component.ts
@@ -1,9 +1,9 @@
 import {Component, OnInit} from '@angular/core';
 import {FormArray, FormBuilder, FormGroup, Validators} from "@angular/forms";
-import {ClassLevel} from "../../../../models/dto/classlevel.model";
+import {ClassLevel} from "../../../../models/dto/class-level.model";
 import {NgbActiveModal} from "@ng-bootstrap/ng-bootstrap";
 import {ClassLevelSubService} from "../../../../services/class-level-sub.service";
-import {ClassLevelSub} from "../../../../models/dto/classlevelsub.model";
+import {ClassLevelSub} from "../../../../models/dto/class-level-sub.model";
 import {addToMessageService} from "../../../../utils/message-service.util";
 import {MessageService} from "primeng/api";
 import {ClassLevelService} from "../../../../services/class-level.service";
@@ -32,6 +32,11 @@ export class SaveClassComponent implements OnInit {
     return this.classForm.get('classLevels') as FormArray;
   }
 
+  ngOnInit(): void {
+    this.setupClassForm(this.classLevel);
+    this.setClassLevelSection(this.classLevel.section_id);
+  }
+
   setClassLevelSection(sectionId: number) {
     this.sectionService.getSectionById(sectionId).subscribe({
       next: (section: Section) => {
@@ -41,11 +46,6 @@ export class SaveClassComponent implements OnInit {
         console.log(err);
       }
     });
-  }
-
-  ngOnInit(): void {
-    this.setupClassForm(this.classLevel);
-    this.setClassLevelSection(this.classLevel.section_id);
   }
 
   resetClassLevel(sectionId: number) {
@@ -60,6 +60,10 @@ export class SaveClassComponent implements OnInit {
       classLevels: this.fb.array([])
     });
 
+    this.loadClassLevelSubs(classLevel);
+  }
+
+  loadClassLevelSubs(classLevel: ClassLevel): void {
     if (classLevel.id > 0) {
       this.classLevelSubService.getAllClassLevelSubsByClassLevelId(classLevel.id).subscribe({
         next: (classLevelSubs) => {
@@ -71,7 +75,7 @@ export class SaveClassComponent implements OnInit {
           });
         },
         error: () => addToMessageService(this.msg, 'warn', 'Warning!', 'Unable to retrieve class level subs')
-      })
+      });
     }
   }
 
@@ -115,7 +119,10 @@ export class SaveClassComponent implements OnInit {
     const confirmDelete = confirm(`Are you sure you want to delete this class level sub: ${classLevelSub.name}`)
     if (confirmDelete) {
       this.classLevelSubService.deleteClassLevelSub(classLevelSub.id).subscribe({
-        next: () => addToMessageService(this.msg, 'warn', 'Delete successful', `Class level sub ${classLevelSub.name} has been deleted`),
+        next: () => {
+          addToMessageService(this.msg, 'warn', 'Delete successful', `Class level sub ${classLevelSub.name} has been deleted`)
+          this.loadClassLevelSubs(this.classLevel);
+        },
         error: (err) => addToMessageService(this.msg, 'error', 'Delete failed', err.message)
       });
     }
